% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facileDataSetList-module.R
\name{facileDataSetSelectServer}
\alias{facileDataSetSelectServer}
\title{Lists the FacileDataSet objects in a parent directory for a user to choose}
\usage{
facileDataSetSelectServer(id, datadir, metafn = "meta.yaml", ...)
}
\arguments{
\item{id}{the ID of the module}

\item{datadir}{the director that holds the FacileDataSet directories}
}
\description{
Enumerates the available dataset objects in \code{datadir} once upon invocation.
A \code{meta.yaml} file in \code{datadir} will tell it how to present (name) and group
the datasets from a dropdown select.
}
