% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Launches the FacileDashboard app}
\usage{
run(
  datadir = "~/workspace/facilebio/data",
  config = NULL,
  user = Sys.getenv("USER"),
  app_title = "FacileDashboard",
  gdb = NULL,
  gdb_idtype = "entrez",
  ...
)
}
\description{
Launches the FacileDashboard app
}
